+++ internal/jimmtest/jimm_mock.go
@@ -13,11 +13,11 @@ import (
 	"github.com/juju/version"
 
 	"github.com/canonical/jimm/api/params"
+	"github.com/canonical/jimm/internal/common/pagination"
 	"github.com/canonical/jimm/internal/db"
 	"github.com/canonical/jimm/internal/dbmodel"
 	"github.com/canonical/jimm/internal/errors"
 	"github.com/canonical/jimm/internal/jimm"
-	"github.com/canonical/jimm/internal/jimm/common"
 	jimmcreds "github.com/canonical/jimm/internal/jimm/credentials"
 	"github.com/canonical/jimm/internal/openfga"
 	ofganames "github.com/canonical/jimm/internal/openfga/names"
@@ -80,7 +80,7 @@ type JIMM struct {
 	InitiateInternalMigration_         func(ctx context.Context, user *openfga.User, modelTag names.ModelTag, targetController string) (jujuparams.InitiateMigrationResult, error)
 	ListApplicationOffers_             func(ctx context.Context, user *openfga.User, filters ...jujuparams.OfferFilter) ([]jujuparams.ApplicationOfferAdminDetailsV5, error)
 	ListControllers_                   func(ctx context.Context, user *openfga.User) ([]dbmodel.Controller, error)
-	ListGroups_                        func(ctx context.Context, user *openfga.User, filter common.LimitOffsetPagination) ([]dbmodel.GroupEntry, error)
+	ListGroups_                        func(ctx context.Context, user *openfga.User, filter pagination.LimitOffsetPagination) ([]dbmodel.GroupEntry, error)
 	ModelDefaultsForCloud_             func(ctx context.Context, user *dbmodel.Identity, cloudTag names.CloudTag) (jujuparams.ModelDefaultsResult, error)
 	ModelInfo_                         func(ctx context.Context, u *openfga.User, mt names.ModelTag) (*jujuparams.ModelInfo, error)
 	ModelStatus_                       func(ctx context.Context, u *openfga.User, mt names.ModelTag) (*jujuparams.ModelStatus, error)
@@ -415,7 +415,7 @@ func (j *JIMM) ListControllers(ctx context.Context, user *openfga.User) ([]dbmod
 	}
 	return j.ListControllers_(ctx, user)
 }
-func (j *JIMM) ListGroups(ctx context.Context, user *openfga.User, filters common.LimitOffsetPagination) ([]dbmodel.GroupEntry, error) {
+func (j *JIMM) ListGroups(ctx context.Context, user *openfga.User, filters pagination.LimitOffsetPagination) ([]dbmodel.GroupEntry, error) {
 	if j.ListGroups_ == nil {
 		return nil, errors.E(errors.CodeNotImplemented)
 	}
